moved all fixtures to conftest.py
