+++ kedro-datasets/tests/databricks/test_base_table_dataset.py
@@ -3,177 +3,10 @@ import os
 import pandas as pd
 import pytest
 from kedro.io.core import DatasetError, Version, VersionNotFoundError
-from pyspark.sql import DataFrame, SparkSession
+from pyspark.sql import DataFrame
 from pyspark.sql.types import IntegerType, StringType, StructField, StructType
 
 from kedro_datasets.databricks._base_table_dataset import BaseTableDataset
-
-
-@pytest.fixture
-def sample_spark_df(spark_session: SparkSession):
-    schema = StructType(
-        [
-            StructField("name", StringType(), True),
-            StructField("age", IntegerType(), True),
-        ]
-    )
-
-    data = [("Alex", 31), ("Bob", 12), ("Clarke", 65), ("Dave", 29)]
-
-    return spark_session.createDataFrame(data, schema)
-
-
-@pytest.fixture
-def upsert_spark_df(spark_session: SparkSession):
-    schema = StructType(
-        [
-            StructField("name", StringType(), True),
-            StructField("age", IntegerType(), True),
-        ]
-    )
-
-    data = [("Alex", 32), ("Evan", 23)]
-
-    return spark_session.createDataFrame(data, schema)
-
-
-@pytest.fixture
-def mismatched_upsert_spark_df(spark_session: SparkSession):
-    schema = StructType(
-        [
-            StructField("name", StringType(), True),
-            StructField("age", IntegerType(), True),
-            StructField("height", IntegerType(), True),
-        ]
-    )
-
-    data = [("Alex", 32, 174), ("Evan", 23, 166)]
-
-    return spark_session.createDataFrame(data, schema)
-
-
-@pytest.fixture
-def subset_spark_df(spark_session: SparkSession):
-    schema = StructType(
-        [
-            StructField("name", StringType(), True),
-            StructField("age", IntegerType(), True),
-            StructField("height", IntegerType(), True),
-        ]
-    )
-
-    data = [("Alex", 32, 174), ("Evan", 23, 166)]
-
-    return spark_session.createDataFrame(data, schema)
-
-
-@pytest.fixture
-def subset_pandas_df():
-    return pd.DataFrame(
-        {"name": ["Alex", "Evan"], "age": [32, 23], "height": [174, 166]}
-    )
-
-
-@pytest.fixture
-def subset_expected_df(spark_session: SparkSession):
-    schema = StructType(
-        [
-            StructField("name", StringType(), True),
-            StructField("age", IntegerType(), True),
-        ]
-    )
-
-    data = [("Alex", 32), ("Evan", 23)]
-
-    return spark_session.createDataFrame(data, schema)
-
-
-@pytest.fixture
-def sample_pandas_df():
-    return pd.DataFrame(
-        {"name": ["Alex", "Bob", "Clarke", "Dave"], "age": [31, 12, 65, 29]}
-    )
-
-
-@pytest.fixture
-def append_spark_df(spark_session: SparkSession):
-    schema = StructType(
-        [
-            StructField("name", StringType(), True),
-            StructField("age", IntegerType(), True),
-        ]
-    )
-
-    data = [("Evan", 23), ("Frank", 13)]
-
-    return spark_session.createDataFrame(data, schema)
-
-
-@pytest.fixture
-def expected_append_spark_df(spark_session: SparkSession):
-    schema = StructType(
-        [
-            StructField("name", StringType(), True),
-            StructField("age", IntegerType(), True),
-        ]
-    )
-
-    data = [
-        ("Alex", 31),
-        ("Bob", 12),
-        ("Clarke", 65),
-        ("Dave", 29),
-        ("Evan", 23),
-        ("Frank", 13),
-    ]
-
-    return spark_session.createDataFrame(data, schema)
-
-
-@pytest.fixture
-def expected_upsert_spark_df(spark_session: SparkSession):
-    schema = StructType(
-        [
-            StructField("name", StringType(), True),
-            StructField("age", IntegerType(), True),
-        ]
-    )
-
-    data = [
-        ("Alex", 32),
-        ("Bob", 12),
-        ("Clarke", 65),
-        ("Dave", 29),
-        ("Evan", 23),
-    ]
-
-    return spark_session.createDataFrame(data, schema)
-
-
-@pytest.fixture
-def expected_upsert_multiple_primary_spark_df(spark_session: SparkSession):
-    schema = StructType(
-        [
-            StructField("name", StringType(), True),
-            StructField("age", IntegerType(), True),
-        ]
-    )
-
-    data = [
-        ("Alex", 31),
-        ("Alex", 32),
-        ("Bob", 12),
-        ("Clarke", 65),
-        ("Dave", 29),
-        ("Evan", 23),
-    ]
-
-    return spark_session.createDataFrame(data, schema)
-
-
-@pytest.fixture
-def external_location():
-    return os.environ.get("DATABRICKS_EXTERNAL_LOCATION")
 
 
 class TestBaseTableDataset:
